+++ Magic/CHANGELOG.md
@@ -15,6 +15,7 @@
  - Fix for mob triggers not properly re-triggering
  - Fix incorrect data type for HideFlags, which could cause issues with protocol plugins (e.g. Protocolize)
  - Add "mount" property to mobs, which can be another mob to use as a mount
+ - Add SetSpawn action
 
 # 7.9.8